Set Socket to only use websocket protocol
This was the root to a lot of issues. If the client refreshed there was a risk that a socket connection was created but then forgotten by the client due to the refresh. If the refresh happened fast enough, the client would not send a disconnect event to the server. Keeping the connection alive forever. This later on resulted in the same event being sent multiple times.