Add function getGameUser
getGameUser will get the user from the game object instead of the top level user object.

We should never make changes to the root user object, since the response is built from user.game

Normally this is not an issue since most changes are done on game.users, but when trying to do changes to "currentUser", One should user gameUser(user)

A better fix for this is probably needed.