Add getter and setter on User Entity
This will make sure to sync the state between the "root" user and "gameUser" at all time, this in turn will abstract away the ambiguity in the differance between the two.

When implementing the user entity, using the getter and setter will result in a more expected behaviour.

To make it clear what prop to use the cards prop has been renamed to player_cards, to reflect that it is the db relation, where as the getter and setter will be called cards.