Refactor and bugfix of get uniqueCards
There was a lot of issues in this code creating a lot of strange bugs, this removes complexity, and is imo a easier flow to follow.

This also solves the bug where the user would not be asigned correct amoutn of cards. This due to some unexpected behaviout of how the new cards where saved to db.