refactor socket lib to not import global types
With this refactor we isolate the "socket lib". Removing all it's dependencies of types outside of the lib module.

The goal of this commit is to make the socket lbi more isolated and stand alone.

To achive the same behaviour as before we instead rely on the implementation of the lib sending in the correct arguments.

For example the creater of the SocketServer will have to send in all middlewares