return the saved instance of new wcr
Since we are setting wcr.user typeorm will add the `wcr.user_id_fk` value to the row on save. If we do not use the wcr instance returned from .save(). The wcr will be missing user_id_fk column. Since it was never explicitly set on the in-memory instance.

This is probably the culprit to many nasty errors including #158 turn reset bug. And maybe duplicate white card bug.